Refactor: Limpando controller de Login parte 2. Controller poderá no máximo chamar uma service.
